+++ src/RcppExports.cpp
@@ -246,6 +246,28 @@ BEGIN_RCPP
     return R_NilValue;
 END_RCPP
 }
+// C_match_instances
+Rcpp::List C_match_instances(Rcpp::DataFrame df);
+RcppExport SEXP _arbor_C_match_instances(SEXP dfSEXP) {
+BEGIN_RCPP
+    Rcpp::RObject rcpp_result_gen;
+    Rcpp::traits::input_parameter< Rcpp::DataFrame >::type df(dfSEXP);
+    rcpp_result_gen = Rcpp::wrap(C_match_instances(df));
+    return rcpp_result_gen;
+END_RCPP
+}
+// C_merge_instances
+void C_merge_instances(Rcpp::List match_result, Rcpp::DataFrame df, int min_support, double min_weight_ratio);
+RcppExport SEXP _arbor_C_merge_instances(SEXP match_resultSEXP, SEXP dfSEXP, SEXP min_supportSEXP, SEXP min_weight_ratioSEXP) {
+BEGIN_RCPP
+    Rcpp::traits::input_parameter< Rcpp::List >::type match_result(match_resultSEXP);
+    Rcpp::traits::input_parameter< Rcpp::DataFrame >::type df(dfSEXP);
+    Rcpp::traits::input_parameter< int >::type min_support(min_supportSEXP);
+    Rcpp::traits::input_parameter< double >::type min_weight_ratio(min_weight_ratioSEXP);
+    C_merge_instances(match_result, df, min_support, min_weight_ratio);
+    return R_NilValue;
+END_RCPP
+}
 // generate_cage_cpp
 Rcpp::DataFrame generate_cage_cpp(Rcpp::DataFrame circles, double decimation);
 RcppExport SEXP _arbor_generate_cage_cpp(SEXP circlesSEXP, SEXP decimationSEXP) {
@@ -531,6 +553,8 @@ static const R_CallMethodDef CallEntries[] = {
     {"_arbor_accumulate_passages_old", (DL_FUNC) &_arbor_accumulate_passages_old, 4},
     {"_arbor_find_closest_node", (DL_FUNC) &_arbor_find_closest_node, 2},
     {"_arbor_C_fix_small_isolated_low_clusters", (DL_FUNC) &_arbor_C_fix_small_isolated_low_clusters, 3},
+    {"_arbor_C_match_instances", (DL_FUNC) &_arbor_C_match_instances, 1},
+    {"_arbor_C_merge_instances", (DL_FUNC) &_arbor_C_merge_instances, 4},
     {"_arbor_generate_cage_cpp", (DL_FUNC) &_arbor_generate_cage_cpp, 2},
     {"_arbor_detect_tree_circles_cpp", (DL_FUNC) &_arbor_detect_tree_circles_cpp, 6},
     {"_arbor_qsm_cpp", (DL_FUNC) &_arbor_qsm_cpp, 2},